move imstat to bokproc; _finish needs to called at completion
